+++ src/providers/MeetingProvider/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2020-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { useContext, useState, createContext, ReactNode } from 'react';
+import React, { useContext, useState, createContext } from 'react';
 import { Logger, LogLevel, VideoDownlinkBandwidthPolicy } from 'amazon-chime-sdk-js';
 
 import MeetingManager from './MeetingManager';
@@ -36,10 +36,6 @@ interface Props {
    * Check `meetingManager` prop documentation for more information.
    */
   meetingManager?: MeetingManager;
-  /**
-   * Whether or not to enable Amazon Chime SDK for JavaScript meeting events.
-   */
-  enableMeetingEvents?: boolean;
 }
 
 export const MeetingContext = createContext<MeetingManager | null>(null);
@@ -51,47 +47,33 @@ export const MeetingProvider: React.FC<Props> = ({
   logger,
   videoDownlinkBandwidthPolicy,
   meetingManager: meetingManagerProp,
-  enableMeetingEvents = false,
   children,
 }) => {
   const [meetingManager] = useState(
     () => meetingManagerProp || new MeetingManager({ logLevel, postLogConfig, simulcastEnabled, logger, videoDownlinkBandwidthPolicy })
   );
 
-  const wrapProviders = (children: ReactNode) => 
-    <AudioVideoProvider>
-      <DevicesProvider>
-        <RosterProvider>
-          <RemoteVideoTileProvider>
-            <LocalVideoProvider>
-              <LocalAudioOutputProvider>
-                <ContentShareProvider>
-                  <FeaturedVideoTileProvider>
-                    {children}
-                  </FeaturedVideoTileProvider>
-                </ContentShareProvider>
-              </LocalAudioOutputProvider>
-            </LocalVideoProvider>
-          </RemoteVideoTileProvider>
-        </RosterProvider>
-      </DevicesProvider>
-    </AudioVideoProvider>;
-
-  const wrapMeetingEventProvider = (children: ReactNode) => {
-    const wrappedProviders = wrapProviders(children);
-    if (enableMeetingEvents) {
-      return (
-        <MeetingEventProvider>
-          {wrappedProviders}
-        </MeetingEventProvider>
-      );
-    }
-    return wrappedProviders;
-  }
-
   return (
     <MeetingContext.Provider value={meetingManager}>
-      {wrapMeetingEventProvider(children)}
+      <MeetingEventProvider>
+        <AudioVideoProvider>
+          <DevicesProvider>
+            <RosterProvider>
+              <RemoteVideoTileProvider>
+                <LocalVideoProvider>
+                  <LocalAudioOutputProvider>
+                    <ContentShareProvider>
+                      <FeaturedVideoTileProvider>
+                        {children}
+                      </FeaturedVideoTileProvider>
+                    </ContentShareProvider>
+                  </LocalAudioOutputProvider>
+                </LocalVideoProvider>
+              </RemoteVideoTileProvider>
+            </RosterProvider>
+          </DevicesProvider>
+        </AudioVideoProvider>
+      </MeetingEventProvider>
     </MeetingContext.Provider>
   );
 };